Hand roll algolia docsearch plugin
