+++ package.json
@@ -79,7 +79,6 @@
     "gatsby": "2.27.5",
     "gatsby-image": "2.6.0",
     "gatsby-node-helpers": "0.3.0",
-    "gatsby-plugin-algolia-docsearch": "1.0.5",
     "gatsby-plugin-google-gtag": "2.3.0",
     "gatsby-plugin-manifest": "2.7.0",
     "gatsby-plugin-mdx": "1.5.0",
