+++ plugins/gatsby-algolia-docsearch/package.json
@@ -1,0 +1,5 @@
+{
+  "private": true,
+  "name": "gatsby-algolia-docsearch",
+  "version": "0.0.1"
+}
