+++ src/layouts/Root/components/Header.tsx
@@ -6,18 +6,16 @@
  */
 
 import React, { useState, HTMLAttributes, useRef, useEffect } from 'react';
-import styled, { css, DefaultTheme, ThemeProps } from 'styled-components';
-import { math } from 'polished';
+import styled, { css } from 'styled-components';
 import { Link } from 'gatsby';
-import { getColor, mediaQuery, menuStyles, PALETTE } from '@zendeskgarden/react-theming';
+import { getColor, mediaQuery, PALETTE } from '@zendeskgarden/react-theming';
 import { IconButton } from '@zendeskgarden/react-buttons';
-import { MediaInput } from '@zendeskgarden/react-forms';
 import { ReactComponent as OverflowVerticalStroke } from '@zendeskgarden/svg-icons/src/16/overflow-vertical-stroke.svg';
-import { ReactComponent as SearchStroke } from '@zendeskgarden/svg-icons/src/16/search-stroke.svg';
 import { ReactComponent as CloseStroke } from '@zendeskgarden/svg-icons/src/16/x-stroke.svg';
 import { ReactComponent as GardenIcon } from '@zendeskgarden/svg-icons/src/26/garden.svg';
 import { ReactComponent as GardenWordmark } from '@zendeskgarden/svg-icons/src/26/wordmark-garden.svg';
 import MaxWidthLayout from 'layouts/MaxWidth';
+import { SearchInput } from './SearchInput';
 import { StyledNavigationLink } from './StyledNavigationLink';
 
 const StyledDesktopNavItem = styled.div`
@@ -94,158 +92,6 @@ const Logo: React.FC = () => (
       </div>
     </Link>
   </div>
-);
-
-const searchStyles = (props: ThemeProps<DefaultTheme>) => {
-  const theme = props.theme;
-  const positionTop = math(`${theme.space.base * 3.5} + ${theme.borderWidths.sm}`);
-  const positionRight = math(`${theme.space.base * -3} - ${theme.borderWidths.sm}`);
-  const highlightBackgroundColor = getColor('primaryHue', 800, theme, 0.08);
-  const hoverBackgroundColor = getColor('primaryHue', 600, theme, 0.08);
-  const hoverSeparatorColor = getColor('neutralHue', 250, theme);
-  const metaColor = getColor('neutralHue', 600, theme);
-  const separatorColor = getColor('neutralHue', 200, theme);
-
-  return css`
-    ${p =>
-      menuStyles('bottom', {
-        animationModifier: '.algolia-autocomplete',
-        childSelector: ' .ds-dropdown-menu',
-        theme: p.theme
-      })}
-
-    /* stylelint-disable
-        declaration-no-important,
-        max-nesting-depth,
-        selector-max-specificity,
-        selector-max-compound-selectors */
-
-    & .ds-dropdown-menu {
-      top: calc(100% + ${positionTop}) !important;
-      right: ${positionRight} !important;
-      line-height: ${theme.lineHeights.md};
-
-      &::before {
-        display: none;
-      }
-
-      & [class^='ds-dataset-'] {
-        border: none;
-        padding: 0;
-
-        & .ds-suggestions {
-          margin-top: ${theme.space.xxs};
-        }
-
-        & .algolia-docsearch-suggestion {
-          padding: 0;
-          color: inherit;
-
-          & .algolia-docsearch-suggestion--highlight {
-            background-color: ${highlightBackgroundColor};
-            padding: 0;
-            color: inherit;
-          }
-
-          & .algolia-docsearch-suggestion--category-header {
-            margin: 0;
-            margin-bottom: ${theme.space.xxs};
-            border-color: ${separatorColor};
-            padding: ${theme.space.xs} ${theme.space.sm} ${theme.space.sm};
-            color: inherit;
-            font-weight: ${theme.fontWeights.semibold};
-
-            & .algolia-docsearch-suggestion--highlight {
-              box-shadow: none;
-              background-color: transparent;
-            }
-          }
-
-          & .algolia-docsearch-suggestion--subcategory-column {
-            padding: 0;
-            padding-right: ${theme.space.sm};
-            color: ${metaColor};
-            font-size: inherit;
-
-            & .algolia-docsearch-suggestion--highlight {
-              background-color: transparent;
-            }
-          }
-
-          & .algolia-docsearch-suggestion--content {
-            padding: 0;
-            padding-left: ${theme.space.sm};
-          }
-
-          & .algolia-docsearch-suggestion--text {
-            line-height: ${theme.lineHeights.sm};
-            color: ${metaColor};
-            font-size: ${theme.fontSizes.sm};
-
-            & .algolia-docsearch-suggestion--highlight {
-              box-shadow: none;
-              background-color: transparent;
-              text-decoration: underline;
-            }
-          }
-
-          & .algolia-docsearch-suggestion--title {
-            margin: 0;
-            font-size: inherit;
-            font-weight: ${theme.fontWeights.regular};
-          }
-
-          & .algolia-docsearch-suggestion--wrapper {
-            padding: ${theme.space.xs} ${theme.space.base * 9}px;
-
-            & > *::before {
-              background-color: ${separatorColor};
-            }
-          }
-        }
-
-        & .ds-cursor {
-          & .algolia-docsearch-suggestion--wrapper {
-            background-color: ${hoverBackgroundColor};
-
-            & > *::before {
-              background-color: ${hoverSeparatorColor};
-            }
-
-            & .algolia-docsearch-suggestion--content {
-              background-color: transparent;
-            }
-          }
-        }
-
-        & .algolia-docsearch-footer {
-          margin: ${theme.space.xs} ${theme.space.sm};
-          width: 120px;
-        }
-      }
-    }
-  `;
-};
-
-const SearchInput = React.forwardRef<HTMLInputElement, HTMLAttributes<HTMLInputElement>>(
-  (props, ref) => (
-    <span
-      css={css`
-        & .algolia-autocomplete {
-          ${p => searchStyles(p)}
-        }
-      `}
-    >
-      <MediaInput
-        start={<SearchStroke />}
-        placeholder="Search…"
-        aria-label="Search"
-        ref={ref}
-        wrapperProps={{ style: { overflow: 'visible' } }}
-        {...props}
-      />
-    </span>
-  )
 );
 
 const MobileSearch = React.forwardRef<HTMLInputElement, HTMLAttributes<HTMLDivElement>>(
